requirements.txt
beautifulsoup4>=4.6.3,<5
lxml>=4.2.4,<5
Pillow>=5.2.0,<6
tensorflow
requests<3,>=2.21.0
python-chess
pyautogui
pyscreenshot

## ONLY IF YOU ARE USING LEGIT MODE
pyclick